slimmed down PredictionPageService.java a bit, filtering still in
